pyproject.toml
[project]
name = "langraph-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.7",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.1",
]
